[Chore] [GGFE-153] 확성기 등록 모달 컴포넌트 이름 변경
